+++ plugins/nf-validation/src/test/nextflow/validation/SamplesheetConverterTest.groovy
@@ -80,7 +80,6 @@ class SamplesheetConverterTest extends Dsl2Spec{
         stdout.contains("[[string1:value, string2:value, integer1:5, integer2:5, boolean1:true, boolean2:true], string1, 25, false, [], [], [], [], [], itDoesExist]")
         stdout.contains("[[string1:dependentRequired, string2:dependentRequired, integer1:10, integer2:10, boolean1:true, boolean2:true], string1, 25, false, [], [], [], unique2, 1, itDoesExist]")
         stdout.contains("[[string1:extraField, string2:extraField, integer1:10, integer2:10, boolean1:true, boolean2:true], string1, 25, false, ${this.getRootString()}/src/testResources/test.txt, ${this.getRootString()}/src/testResources/testDir, ${this.getRootString()}/src/testResources/testDir, unique3, 1, itDoesExist]" as String)
-        stdout.contains("class nextflow.validation.ImmutableMap")
     }
 
         def 'should work fine - TSV' () {
@@ -133,120 +132,6 @@ class SamplesheetConverterTest extends Dsl2Spec{
         stdout.contains("[[string1:value, string2:value, integer1:5, integer2:5, boolean1:true, boolean2:true], string1, 25, false, [], [], [], [], [], itDoesExist]")
         stdout.contains("[[string1:dependentRequired, string2:dependentRequired, integer1:10, integer2:10, boolean1:true, boolean2:true], string1, 25, false, [], [], [], unique2, 1, itDoesExist]")
         stdout.contains("[[string1:extraField, string2:extraField, integer1:10, integer2:10, boolean1:true, boolean2:true], string1, 25, false, ${this.getRootString()}/src/testResources/test.txt, ${this.getRootString()}/src/testResources/testDir, ${this.getRootString()}/src/testResources/testDir, unique3, 1, itDoesExist]" as String)
-    }
-
-    def 'immutable meta' () {
-        given:
-        def SCRIPT_TEXT = '''
-            include { fromSamplesheet } from 'plugin/nf-validation'
-
-            params.input = 'src/testResources/correct.csv'
-
-            workflow {
-                Channel.fromSamplesheet("input", schema_filename:"src/testResources/nextflow_schema_with_samplesheet_converter.json")
-                    .map { 
-                        println(it[0].getClass())
-                        new_meta = it[0] + ["id":"test"]
-                        println(new_meta.getClass())
-                    }
-            }
-        '''
-
-        when:
-        dsl_eval(SCRIPT_TEXT)
-        def stdout = capture
-                .toString()
-                .readLines()
-
-        then:
-        noExceptionThrown() 
-        stdout.contains("class nextflow.validation.ImmutableMap")
-        !stdout.contains("class java.util.LinkedHashMap")
-    }
-
-    def 'mutable meta' () {
-        given:
-        def SCRIPT_TEXT = '''
-            include { fromSamplesheet } from 'plugin/nf-validation'
-
-            params.input = 'src/testResources/correct.csv'
-
-            workflow {
-                Channel.fromSamplesheet("input", schema_filename:"src/testResources/nextflow_schema_with_samplesheet_converter.json", immutable_meta:false)
-                    .map { 
-                        println(it[0].getClass())
-                        new_meta = it[0] + ["id":"test"]
-                        println(new_meta.getClass())
-                    }
-            }
-        '''
-
-        when:
-        dsl_eval(SCRIPT_TEXT)
-        def stdout = capture
-                .toString()
-                .readLines()
-
-        then:
-        noExceptionThrown() 
-        !stdout.contains("class nextflow.validation.ImmutableMap")
-        stdout.contains("class java.util.LinkedHashMap")
-    }
-
-    def 'mutable meta - param' () {
-        given:
-        def SCRIPT_TEXT = '''
-            include { fromSamplesheet } from 'plugin/nf-validation'
-
-            params.validationImmutableMeta = false
-            params.input = 'src/testResources/correct.csv'
-
-            workflow {
-                Channel.fromSamplesheet("input", schema_filename:"src/testResources/nextflow_schema_with_samplesheet_converter.json")
-                    .map { 
-                        println(it[0].getClass())
-                        new_meta = it[0] + ["id":"test"]
-                        println(new_meta.getClass())
-                    }
-            }
-        '''
-
-        when:
-        dsl_eval(SCRIPT_TEXT)
-        def stdout = capture
-                .toString()
-                .readLines()
-
-        then:
-        noExceptionThrown() 
-        !stdout.contains("class nextflow.validation.ImmutableMap")
-        stdout.contains("class java.util.LinkedHashMap")
-    }
-
-    def 'immutable meta - put' () {
-        given:
-        def SCRIPT_TEXT = '''
-            include { fromSamplesheet } from 'plugin/nf-validation'
-
-            params.input = 'src/testResources/correct.csv'
-
-            workflow {
-                Channel.fromSamplesheet("input", schema_filename:"src/testResources/nextflow_schema_with_samplesheet_converter.json")
-                    .map { 
-                        it[0].string1 = "unwanted change"
-                    }
-            }
-        '''
-
-        when:
-        dsl_eval(SCRIPT_TEXT)
-        def stdout = capture
-                .toString()
-                .readLines()
-
-        then:
-        def error = thrown(nextflow.exception.AbortRunException)
-        error != null
     }
 
     def 'extra field' () {